requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via environ-config
environ-config==24.1.0
    # via -r requirements.in
tenacity==9.1.2
    # via -r requirements.in
